requirements.txt
Flask
gunicorn
tensorflow
numpy
pandas
tensorflow-recommenders
mysql-connector-python
requests